d3/pyproject.toml
[project]
name = "d3-mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyyaml>=6.0.3",
]
